Make `run_namd_min.sh` work with current CentOS6
NAMD modules on the cluster don't seem to load for me. Instead, it's really easy to download and untar NAMD. The script simply uses `namd2` located wherever NAMD was untar'ed to. Furthermore, some of the sed commands needed updating. Maybe with the updating of namd2 version, the way file locations are specified changed. I found that every path needs to be relative to wherever the .conf file is located.